remove unsupported features
