+++ bin/setup.py
@@ -17,9 +17,7 @@ setuptools.setup(
     packages=setuptools.find_packages(),
     entry_points = {
         'console_scripts' : [
-            'tprism = tprism_module.torch_tprism:main',
-            'tprism_torch = tprism_module.torch_tprism:main',
-            'tprism_tf = tprism_module.tf_tprism:main',
+            'tprism = tprism.torch_tprism:main',
             ],
     },
     classifiers=[
